backup_pitr: tar tablespaces in separate files
Each tablespaces defined in the cluster are tar'ed in a file with its
name, this ensure that archives are recognizable for human beings.  No
conflits are possible since the name are taken from PostgreSQL and are
unique.

Also optimize the code to avoid repetition in test between local and
remote storage.