archive_xlog: make archiving fail more strictly
Before this patch, only a failure on all targets return an error to
PostgreSQL. This is not good when a slave goes down then up, because it
would get a hole in the WAL files chain, the admin would less easily
find what happened.

Now, if a copy fails, it is recorded and the process continues on the
other targets. When done, the script give return code 1 to PostgreSQL
to make the master retry later.