fix backup selection by using stop time in UTC
A backup is cohorent only after pg_stop_backup() has been issued. Thus
the target date used for restore must be after the stop time of the
backup label when selecting the base backu to restore

To avoid any problems with time zones, use Unix timestamps when
comparing dates. The timestamp of the stop time is then stored in
backup_dir/backup_timestamp to be usable when restoring. The target date
specified on the command line on restore is now check to have a proper
format with an optional time zone offset on 4 digits with a +/- sign