Allow usage of hostnames to specify hosts
All script requiring to resolv hostnames or IP address now use a big
hack because function like gethostbyname do not exists as single
binaries to let us resolv hostnames to IP addresses honoring name
service switch. So we use ping (and ping6) to resolve them. The only
limitation is the lack of consistency between GNU and *BSD on the ping6
output.