Typo fixes
Cherry picked off the top of the other patches which are to follow.
Nothing here changes the functional operation of any code, and they
should mostly be self-evident, except perhaps:

There is no $backup_dir defined in list_pitr (anymore?).

The "transfering PGDATA" message wasn't updated for the code which
is actually transferring the tablespaces (and the english spelling
really is always transferring everywhere).

The docs for ARCHIVE_HOST in the sample config were describing the
behaviour of an earlier version?  In the current code it is not enough
to leave that empty for local archives, ARCHIVE_LOCAL must be set too.