Add manpages for the 'user facing' commands
The library actions are documented in the pitrery.1 page.

The INSTALL.md contains a great description of the details of using and
configuring pitrery, but having to find it (and remember that's where
this is) and grep through it for some usage detail that you've forgotten
isn't as convenient as just being able to pull up a manpage quickly.

Install them using the Makefile. Ensure they are installed in the RPM
package. The Debian package will take care of it by itself.