Use bash arrays to build up command line arguments
And factor out the common operations of testing if a library command is
available, adding the remote hostname parameter, and running the command.

Using an array is the only safe way to assemble a command line from a
sequence of arbitrary input strings.  It ensures that we always know
where the intended word boundaries are, and can preserve them regardless
of what characters appear in any of the arguments.

There's a bunch of other places that we'll do this too, and this was
actually one of the last that I'd done in this series of patches, but
it's also one of the most obviously easy to review, so I've popped it
off the stack early in this set of commits now.  Some of it could still
be refactored further, but this is the important improvement we want to
make things simpler and more robust here.