Add qw() for safely quoting remote arguments
While we can use arrays to make assembling local command lines safe, that
doesn't help us for things which will be expanded more than once, such as
the remote arguments to ssh, scp, and remote rsync (which passes them to
ssh).  To protect those we need a slightly different approach.

The qw() function will add an extra layer of shell quoting to the list of
arguments that are passed to it, such that when the resulting string is
parsed again by the remote shell, the original word splitting and quoting
will be correctly preserved.

We can't really do this any other way either, since it's not safe to just
single or double quote the remote arguments, because that will explode if
any of the locally expanded variables also contain those characters.

Ideally, this function probably belongs in a common library file that we
just source everywhere we want to use it, but I'll leave that for later
refactoring too, because there's a bunch of other common functions here
that probably should move there too if it's worth doing that.