purge: add a dry run option
As a single patch, this is a bit more complex than it ideally could have
been, mainly because it started as a patch to remove the creation and
parsing of temp files, replacing them with arrays which were safe for
any arbitrary characters in the strings.  That patch actually greatly
simplified much of the logic here, but it also opened up the possibility
to fairly easily compute the final result without actually removing any
of the existing files, making a dry run option feasible to add -- which
I then spent all the newly found simplicity on actually adding ...

I probably should have done a checkpoint commit before starting on that
since the original change evolved around it even further which makes it
a bit harder to sensibly separate them again now.  So I'm pushing this
all together now as a single 'atomic' change to add the dry run option.

Some things of note, more or less in the order they appear in the patch:

We don't need a tmp_dir at all anymore for this.

Add ssh_target and xlog_ssh_target instead of using the full incantation
'${ssh_user:+$ssh_user@}$target' everywhere we need it.  That just makes
things a little more readable, and reduces the risk of typo'ing it.

Without 'shopt -s nullglob' if there are no matches at all for the glob
then the literal glob string will be returned.  That's pretty much never
what we want to happen anywhere here that we use globs.

Avoid using `ls` to create a list of files to iterate over with a for
loop.  That will fail when the string is word split if any of the files
have whitespace or other shell metacharacters in their names.  On the
local host, where we know the shell is bash, it's better to just expand
the glob directly into an array.  For getting a list of files on the
remote host, where we can't be sure what the default shell is, or if
bash is even installed there, we can use find -print0, then split the
string into an array on the nul characters, since it's not legal in a
filename.

By putting a trailing '/' on the glob, we can make it match only
directories.  If we strip that trailing character off again when we
iterate over the array, we have the same result as `find -type d`.

Use `ssh -n` everywhere that we don't pipe something into it, to avoid
having it accidentally stream anything in stdin to the remote side.
There's a few places we _need_ that, but it's safe and sane to just use
it in all the places we aren't intentionally piping something to it.

Use '--' to avoid accidentally having an argument expanded from a
variable interpreted as a command option.  We know all the places where
that is never what we'd want to happen.

Use qw() to quote any remote argument expanded on the local side, to
ensure it is seen correctly when expanded again on the remote site.

Validate that things we expect are numbers, really are numbers before
using them in arithmetic.  Things often won't go like we might first
hope if they are not.

Use (( )) for arithmetic evaluation, it's better behaved in almost
every (corner) case than [ ] is when dealing with numeric comparison.

Don't bother capturing $? if we don't actually need the return code to
pass on as our own exit code.  To test if a command succeeded we can
just do `if command; then ...`.

The 'what to purge' logic is a bit more convoluted than it possibly needs
to be, partly because the original code caters to the case where a backup
might have a backup_label, but does not have a backup_timestamp ...
This shouldn't normally happen, and there are several alternative things
we could do if it does, but right now I've stayed on the side of being
100% compatible with the previous behaviour.  There might be an argument
for changing some of that in the future - but this commit isn't the place
to make that argument.

There's some opportunistic simplifications like `grep | cut | sed` => awk
and `echo $x | cut -b 1-8` => ${x:0:8}, and replacing grep with [[ ]] as
a pattern or regex match, which don't actually change the previous
behaviour at all, and could probably have been split into a separate
patch here, but that probably wouldn't have made this one easier to
review in any notable way.

Many of the above items are applicable to other code here too, but I will
keep all those changes separate where they aren't directly part of some
other more genuinely notable change.

Edit: while integrating this patch, some bugs have been fixed on ipv6
address escaping and the whole WAL file purge logic has been rewritten
with performance in mind. Since the list of archived WAL files comes
ordered alphabetically, we get them from the oldest to the newest, so we
can just walk the array and stop when we found the oldest one to keep.

Expanding an array to build an list of arguments for rm can lead to an
error from rm finding the list is too long. So just revert to the
previous technic of piping rm commands of each file to a shell, it may
look uglier but it scales.

Finally, improve display of backups to purge and only list the first and
last WAL files that we would remove.