archive_xlog: don't overwrite an existing file
This is in line with the PostgreSQL recommendation to use something like:
'test ! -f $dir/%f && cp %p $dir/%f' as a simplest archive_command to
avoid accidentally destroying existing backups.  In their words:

 The archive command should generally be designed to refuse to overwrite
 any pre-existing archive file.  This is an important safety feature to
 preserve the integrity of your archive in case of administrator error
 (such as sending the output of two different servers to the same archive
 directory).

Which is a distinct possibility here if the wrong pitrery config file or
label is used.  A design that makes handling multiple clusters easy also
makes a mistake like that more easy to make at some point too.

Edit: when integrating this patch, a new parameter ARCHIVE_OVERWRITE has
been added, enabled to "yes" by default, so that the check it not
performed. This means, that while recommended, we do not fully agree
with this. The risk of having badly copied files that blocks archiving
is worse and the overhead of the check over SSH is big. The only real
dangerous issue is an archiving enable on a promoted standby that could
corrupt the files when promoted (before 9.5) but we do not have to take
care of this with PITR. Letting the user choose is a good compromise.

Use a tempfile created securely with mktemp for the compressed WAL file
rather than a file in /tmp with a predictable name (and these days, even
relying on /tmp being the right place for things like that for any given
user or service has become a more uncertain proposition anyway).

Be defensive about quoting variables, for both local and remote use.
Once the rest of this code is updated for that too, it will be safe for
file and path names to contain arbitrary characters including whitespace
and other shell special characters, and they will be preserved, but this
patch only updates the things it was already touching for that.