Support LOG_TIMESTAMP in {archive,restore}_xlog
This is useful when stderr is going to the normal postgres logging
collector rather than syslog, to know when some event occurred if it
isn't bracketed by other recent output from postgres.

Use $* rather than $1 in error() and warn(), mostly for consistency with
the definitions we use for them everywhere else.  We don't actually ever
pass multiple options to those anywhere, but there's no good reason for
them to behave differently in different places either (and this lays the
foundation for them to be moved to a common library file to be sourced
everywhere that uses them if that turns out to be a desirable thing at
a later point in time too).