restore: relax restrictions on the target date format
There really isn't any reason to force the user into using a strict
format for the -d option.  So long as date(1) on their system accepts
what they provide we can massage it into the form(s) we need to set up
the restore operation.  And it can be useful when you are in a hurry
to be able to say `pitrery restore -d '15 minutes ago' without having
to remember strict formats and fiddle with timestamp math and check
timezones etc.

The bulk of this patch is actually converting restore to use arrays
instead of temp files for its data processing, which both significantly
simplifies a lot of the logic and makes it more robust to strings with
difficult characters in them - which in turn is what invited relaxing
the target time handling, so those changes are somewhat intertwined.

There is one exception in this case to the goal of being perfectly
robust no matter how hostile the input string or how silly the typo in
the configuration or command line was - and that is the tblspc_list
file.  It is exported from the cluster with a query using the default
pqsl field separator '|', which means if any of those fields contain a
literal '|', the game is over because psql won't escape them, and even
if it did, the code here assumes that won't ever happen.  And to be
fair, it _probably_ won't, if you put a | in the path of your tablespace
dir, you're just begging for the pain that will cause you, but it does
leave us a small step short of 'perfectly safe' even in a hostile world
full of accidentally stupid and/or fat fingered people.

It would be possible to fix that, we could pass '-z -0' to psql when
that data is exported, and parse it as nul terminated fields, but doing
that wouldn't be backward compatible with existing backups, which every
other change here has taken every care to still be.  If we want to fix
this issue it will take some careful thought and a migration strategy.

Other things of note which might make this easier to review:

We simulate a multi-dimensional array for the tablespace relocations
with four variables, tspc_name, tspc_dir, tspc_oid, and tspc_reloc.
For each row index, they contain the corresponding fields from the
exported tblspc_list that was created at backup time.  We can iterate
over all of them together by index, or individually as an array list.

We extend the check for 'tar' or 'rsync' storage type to also check if
it looks like a tar format backup with a different $compress_suffix.
That seems a bit more helpful to report if the only reason it could
not find the format was because BACKUP_COMPRESS_SUFFIX doesn't currently
match what was used at the time the backup was taken.

Don't use sed where a bash parameter expansion will do.

There's no need to cd to create a symlink, and ln can force remove the
target to avoid a race with a separate call to rm.

Be more defensive about quoting as described for earlier commits.