restore: fixes for check_and_fix_directory()
There's no need to return an exit status here (anymore), there was only
one (remaining?) case where it would be non-zero, and only one place
where the caller actually checked it (which shouldn't normally hit that
case, and definitely won't now that it's quoted defensively).  So just
error() out instead if it happens, because if it does there was a bug
in the calling code - this should be 'impossible' now, and the check is
just an assert of that fact.

mkdir can set the permission of directories it creates, there's no need
to race with a separate call to chmod for that.  If we aren't setting
any special mode bits, this is atomic.

Use `ls -A` to check if the directory is really empty, that will catch
any .dotfiles which might be in it too.  This is about the only case
where ls is better than directly importing a glob, and only because we
don't try to parse the result, we just care if it's empty or not.

Move the permission check into the branch where we have an existing
dir.  We don't need to check that if we created it, because we just
set its permission ourselves when we did.

Be more defensive with quoting.