restore: let rsync use its optimal delete method
If we just use --delete here, then rsync versions prior to 3.0.0 will
default to using --delete-before, the same as this was previously, but
newer rsync versions will instead use --delete-during, which has
basically the same advantage (of freeing up any space possible before
copying new data), except it does it more efficiently, per-directory,
which means the transfer can start more quickly without blocking on a
scan of the full file list and it will use less memory to do it.

I'm also tempted to remove the hard coded "ssh -o Compression=no" here,
since that seems like something which is better placed in a .ssh/config
for the host in question or at least made a configurable option.  On a
fast network with a 'slow' CPU, this option might be an advantage, but
it might not be the best option for a long distance link, at least not
if it's not also used with the rsync -z option.  But I'll leave that
for now since this is a more intrusive change, with no Clearly Right
answer, that would benefit from some discussion of the various options
first.