list: use bash arrays to avoid word splitting trouble
Avoid using `ls` to create a list of files to iterate over with a for
loop.  That will fail when the string is word split if any of the files
have whitespace or other shell metacharacters in their names.  On the
local host, where we know the shell is bash, it's better to just expand
the glob directly into an array.  For getting a list of files on the
remote host, where we can't be sure what the default shell is, or if
bash is even installed there, we can use find -print0, then split the
string into an array on the nul characters, since it's not legal in a
filename.

By putting a trailing '/' on the glob, we can make it match only
directories.  If we strip that trailing character off again when we
iterate over the array, we have the same result as `find -type d`.

We can also replace `awk '{ print $1 }'` with shell code by just
importing the output into an array and accessing the first element.

We can replace the sed prefix stripping with parameter expansion, and
do a more robust check for the (unlikely) case where there is more than
one pgdate.tar.* in the backup directory.