list: make better use of awk
It can read a file directly, we don't need to cat things into it.
The main reason for this patch though was to be more defensive about
the quoting and make it safe for $dir to contain arbitrary characters,
moving where the filename appeared was just a side issue to that.

We probably could rearrange all of this and replace 3 calls to ssh
with just a single connection that collects all the data we need,
since we're basically doing small variations of exactly the same query
each time, so we could just import all the fields like we do in restore
and then process them locally.  For now though, I was mostly interested
in just having it not misbehave awkwardly on unfriendly pathnames.

If we really wanted to go crazy and speed this up, the entire remote
query for each backup directory could probably be done in one giant and
evil ssh command - but it's not clear that is really better than keeping
the logic as similar as possible to the local side handling, and if we
really care about speeding this up, we'd probably get better results
more simply by using the ssh ControlMaster option and multiplexing all
of the remote commands we want to run over a single connection session.
But worry about all that later too.  Maybe ...