backup: ensure the hook variables are always available
The corner case here was that if a POST_BACKUP_COMMAND was set without
also having a PRE_BACKUP_COMMAND, then most of the hook variables would
not actually be set or exported.

We could duplicate that in post_backup_hook(), but it's probably cleaner
to just always export them early.  They're namespaced sufficiently that
it shouldn't conflict with anything else (and if they did that would
still be a problem for anyone who used the PRE_BACKUP_COMMAND.

Drop the `export PITRERY_EXIT_CODE=${PITRERY_EXIT_CODE:-0}` hackery.
It's not clear to me what problem the comment there was really worried
about, but I can only assume that was something from an earlier version
of this code, because the current code always explicitly sets the
PITRERY_EXIT_CODE variable immediately before calling post_backup_hook()
(and nothing touches it afterwards) so that can't be a problem (anymore).
We just need to mark it as being exported.

One other thing that it would be nice to do here, is to make the
POST_BACKUP_COMMAND and PRE_BACKUP_COMMAND variables arrays ...
That way it would be possible to configure and use commands like:

PRE_BACKUP_COMMAND=( 'cmd' '--option="some terrible syntax"' )

which is basically impossible right now in the worst possible cases
since a simple string can't strictly control the word splitting that
will occur.

We could probably even do it in an *almost* backward compatible way,
since if the code executes "${PRE_BACKUP_COMMAND[@]" (with the double
quotes), then both the above and PRE_BACKUP_COMMAND="cmd" would work
correctly - but it would fall over if someone was using a horror
command where they'd tried to explicitly escape the nasties on the
assumption it would be word split as a simple string.

The alternative would be to add a separate command variable, and execute
the array one if it was set, else use the legacy one (and possibly warn
about updating your config).  But this is a change that will need some
further discussion first.