backup: don't use ls to find the previous backup
The `tail -1` makes this case a little less susceptible to trouble than
most of the other uses of parsing `ls` output we've replaced here, but
linebreak characters are legal in filenames, and that would potentially
break this one too.

Which isn't to say I have much sympathy for what happens to people who
deliberately do something as insane as that, but we still shouldn't do
even more unexpected things to people who that might happen to through
some accident or another.