backup: exclude a few more things from the backups
Exclude the restored_config_files directory.  If that is not manually
removed after a restore, before a new backup is taken from the restored
PGDATA dir which includes it, then if that second backup is restored
from, the config files from that backup will end up being placed in a
'restored_config_files/conf' subdirectory rather than having the 'conf'
temporary directory be renamed to restored_config_files when it is
copied into PGDATA.

We could hack around that in restore, by removing restored_config_files
before doing the copy (cp has no clean way to force replacing it), but
there really isn't any good reason for it to be in the backup in the
first place, it will only confuse people who'll find two sets of config
files, the current lot, and the ones from whichever backup was last
restored.

Exclude backup_label.old too.  If the backup_label file is not manually
removed after a restore before a new backup is taken, then the previous
backup_label file will be renamed to this.  There's likewise no good
reason for new backups to continue carrying those around.

Use --delete-excluded when copying with rsync.  As described in the
change to restore, this will use the optimum --delete method depending
on the rsync version (defaulting to --delete-before for older rsync),
but it will also ensure any of the excluded file patterns are removed
from the backup, if they got there through some other means (like the
copy we do when creating hardlinks, if those file were included before
this change was deployed).

And I again question whether "ssh -o Compression=no" should be hardcoded
here too, but that still needs to be discussed before changing it.