backup: sanity check the target directory doesn't already exist
The odds of a collision here are pretty small too, two backups would
have to be done to the same dir, with the same label, and if using the
default directory timestamping, would have to complete within the same
second of each other -- but this is cheap to test and the behaviour if
the directory already exists is sure to cause wondrous confusion, since
it will be placed _inside_ the colliding directory rather than simply
overwriting it (or failing while trying).  At least for the first
collision of some target name.

There is still a race we could lose between checking and actually doing
the mv, but this patch makes it a little smaller, and does protect
against the slightly more probable case where the two backups aren't
really racing, but rather one is already there, either because a clock
was wrong, or because of a timezone change or similar.

There are actually lots of other terrible things that could possibly
rain down on you if you are at risk of this happening (like backups
being stored 'out of order', or having backups from different clusters
all mixed together) - but its still better to warn you if we can.

This does make something of a case for using UTC for the timestamp
based directory names too, since there are a few places still where
we sort backups based on their lexical ordering (rather than the
content of the backup_timestamp file), and local clocks going backward
at the end of daylight saving, or because a standby in a different
timezone was promoted do risk creating confusion there if two backups
are run more often than 'daily' ...   Maybe we want an even stronger
check here, like the previous backup dirname sorting earlier than
this one?

I'm going to leave those questions open for Further Discussion too.