backup: build the psql_command line as an array
This means we don't have to worry about trouble with word splitting of
its arguments, like we might if it was a simple string.

The main remaining catch here is the PITRERY_PSQL variable passed to the
hook commands.  We can't export an array as an array, since arrays are a
bash (and zsh etc.) thing, and the command could be a C executable or
similar, which will only recognise plain environment variables.

That's not a new problem here, it just means pitrery itself might now
work with things that a hook script could have some trouble with, and
what is passed to the hook scripts is still 100% backward compatible.
However a good solution to this is far less likely to remain so ...

There's a few ways we could solve this.  Possibly the easiest and best
would be to change the hooks to be sourced into a subshell rather than
simply executing them as external commands.  That would require them to
also be bash shell snippets, but they could still just be a wrapper to
parse the variables and construct a command line to call some other
program themselves.

Alternatively we could export PITRERY_PSQL_CMD, PITRERY_PSQL_ARG1,
PITRERY_PSQL_ARG2 ... etc.  but that's kind of ugly whichever way you
look at it.

I'm going to leave the best option for that as a matter for further
discussion as well at this point.