Assert we aren't configured for both local and remote backups
In a lot of the code, the 'local' configuration will override use of a
remote host, but ensuring that is always true, both now and forever,
and that there are no unexpected side effects to anywhere that isn't
is a problem we are simply much better of just not having.

So bail out early if the user supplies a confused configuration that
has both local and remote options selected for either the ARCHIVE_ or
BACKUP_ options.  It's ok for one of them to be remote and the other
local, but neither of them should be configured to be both.