Avoid the use of test -a/-o for and/or constructs
They have been marked obsolescent by POSIX because expressions using
them aren't always unambiguous.  This code is already using the
preferred form in most places, so get rid of the last of these now
while I'm auditing everything and be done with it.

And be more defensive with quoting.

Nothing in this commit changes any behaviour in any normally expected
case.