Don't separately test $? for simple success results
Not when we can just test the command directly.

The real reason for pretty much all of this commit is to tighten up the
quoting, but there's no good reason not to simplify things while we are
touching them.

Nothing in this commit changes any behaviour in any normally expected
case (except for the added protection against arbitrary characters).