restore: fix the indent in the tablespace handling loop
We dropped a layer of enclosing scope in the modification to make that
use arrays, but I left the original indent to make the other changes
there easier to review in the diffs.  Now that we've pushed all those
off the stack, we can safely just fix the whitespace too.

There's no code change here, just one less level of indent now.