Don't include the content of pg_replslot in backups
We only exclude the content, not the whole directory, since that saves
us from having to mess about manually recreating it at restore time.

The pg_basebackup tool deliberately omits cloning the replication slots,
which does seem to be the right thing to do in almost every case, at
least with the support for them in postgres 9.4/9.5 anyway.

The rationale for that is probably best described in the words of the
postgres documentation itself:

 It is often a good idea to also omit from the backup dump the files
 within the cluster's pg_replslot/ directory, so that replication slots
 that exist on the master do not become part of the backup. Otherwise,
 the subsequent use of the backup to create a standby may result in
 indefinite retention of WAL files on the standby, and possibly bloat on
 the master if hot standby feedback is enabled, because the clients that
 are using those replication slots will still be connecting to and
 updating the slots on the master, not the standby. Even if the backup is
 only intended for use in creating a new master, copying the replication
 slots isn't expected to be particularly useful, since the contents of
 those slots will likely be badly out of date by the time the new master
 comes on line.

There's a more detailed discussion of the problem(s) with this and the
proposed (possible future) solutions in this thread:

http://www.postgresql.org/message-id/flat/CAB7nPqTBik66XosbXpRoSneQQ9EPPXr0tVpLc4vGLB5hWz0Rkg@mail.gmail.com

A patch with a more functional workaround for the case where you do want
the slots restored as they were at the time of taking a base backup with
pitrery will follow this one.