Record the replication slots that existed at backup time
And provide an easy way to recreate them freshly (without the stale
state that would be included by copying the pg_replslots directory)
after a restore.

We keep a record of them in a similar way to how the tablespace list
is captured, and provide SQL scripts at restore time which can
recreate them in a similar way to how the catalog is updated for
tablespace location changes with postgres prior to 9.2.  This puts
the user in full control of whether to recreate them for any given
restore, and which ones to recreate if only a subset are wanted.