Exclude *.sql files from backups too
This might be a little over-broad, mostly what we want to do with this
change is ensure that stale restore_(in)active_replication_slots.sql
files are not included in a backup, since restoring them if they aren't
overwritten by a new one could lead to confusion and accidents.

The same also applies to the generated update_catalog_tablespaces.sql
though postgres 9.1 has been old for quite a while now, so probably
not all that many people will still hit that one.

We could make this more explicit, but there's probably not many reasons
for people to be storing random .sql files in PGDATA that they are
expecting to be retained in backups ...

The excludes could probably benefit from being factored out now so that
they can just be defined in one place rather than needing to keep four
of them in sync, but that can wait for now.  If there are many more to
add, it's probably worth doing that though.