Wrap bare IPv6 addresses in []
Which is needed to avoid confusing the ':' in them with port or path
delimiters in a lot of remote host references.  This commit adds that
to the list and restore functions, where it was previously missing,
and replaces grep with a [[ ]] pattern match in the rest.

A direct replacement for the previous code would be just [[ $x == *:* ]],
but we can trade some gained simplicity for a stronger test again and now
we only do this if there are no square brackets already in the address
string.  That way users which are already in the habit of wrapping IPv6
literals in this way won't get them wrapped again.

This is still a pretty dumb test for "is this an IPv6 address", but it
should be good enough for anything we expect to be passed in here.