archive_xlog: make error() actually exit
This makes it consistent with the error() function used everywhere else
in this code, and so less likely to be accidentally used wrongly (which
I've already caught myself doing once now, despite knowing it was so).

The only half good reason for it not to be doing that was so we could
propagate the real exit code from commands that failed, but postgres
doesn't care about that, it only wants to know if we returned zero or
non-zero, and the precise exit code isn't all that much use without
knowing exactly what command returned it anyway.

This lets us simplify a lot of this code, which is good, and reduces
the risk of future changes introducing new bugs, which is even better.
Especially since some of the failure paths here aren't likely to see
a lot of actual exercise beyond (hopefully!) being tested when they
are first written.

Edit: in fact postgres do care about the return code of the command.
If it is > 128, the command died on a signal, which make the archiver
process exit. If it is not taken into a slow archive command could leave
a zombie archiver in an HA setup for example, where postgres should die
fast if requested.

So we just rework the error() function to get the exit code from its
second argument. We make it clean the temporary compressed file too.

Conflicts:
	archive_xlog.bash