Be more defensive with quoting
And this should be about the last of the stragglers, the code where
we've had no other useful change to make except to the quoting.

I've erred on the side of being paranoid and conservative here, a few of
these aren't actually a problem in the current code, or could only become
a problem if the code changes or a bug is introduced, but it's easier and
better to just quote all the things that might under some circumstance
need it now than to have to go back and continually re-check the 'safe'
exceptions to that each time later changes are made to the code again.

There's a few places where I've dropped definitely unnecessary quoting,
such as the argument to case statements and simple variable assignments
which don't need it whatever the variable contains, and dropped some
braces where they weren't needed to make a name unambiguous too.

With this last change to the quoting, we should now be able to safely
handle any characters in paths and arguments, aside from the issue of
pipe characters being ambiguous in the database rows exported to the
tblspc_list file (which is also fixable, but probably not in a fully
backward compatible way).