Use rsync --link-dest to create hardlinks
That has been supported by rsync for over 10 years now, so we should be
able to rely on it being widely available.  This should be faster than
manually creating a hardlink tree of everything first, it simplifies the
code significantly, and we no longer have to worry about configuring a
portable HARDLINKER depending on the target system.

In theory we could use this to copy the postgres config files over too,
since they aren't all that likely to change between most backups, but
the amount of space we'd really save with that might not be worth the
extra complication.  It wouldn't be that hard to do if someone cares
enough for it though.