archive_xlog: clean up after a failed attempt to compress
If the compression fails during local archiving, it could leave us with
a stub of a WAL file before returning a failure exit code to postgres.
This means that when postgres later retries to archive it again, we will
see that it already exists and refuse to overwrite it, so we need to
delete it before returning failure.

This isn't a problem for the base backups, because if they fail we will
remove the entire backup directory that was just created.